Split pipeline_client/agent/phases.py into a phases/ package
Decompose the ~1,700-line monolithic phases.py into cohesive per-phase
modules (discovery, image_resolution, issues, finance, refinement,
polling, forecast, voter_resources, iteration) plus a shared_runner
orchestrator and a _common helpers module. Pure refactor: no behavior,
prompt, retry-logic, or model-selection changes.

phases/__init__.py re-exports the full previous public surface so every
existing import site keeps working. Names that tests patch directly on
pipeline_client.agent.phases (_agent_loop, _candidate_source_hints,
_sync_ballotpedia_roster, _ballotpedia_election_lookup,
fetch_kalshi_market_signals) are looked up via a lazy `from . import`
inside the calling function in each submodule, rather than a top-level
import, so monkeypatching the package namespace still takes effect
regardless of which submodule performs the call.
